+++ sources/VS/Panel/src/GUI/GovernorGUI.cpp
@@ -31,10 +31,10 @@ void GovernorGUI::OnPaint(wxPaintEvent &)
    
     dc.DrawCircle(radius, radius, radius);
 
-    float r = radius * 0.6F;
+    float r = static_cast<float>(radius) * 0.6F;
 
-    float x = radius + Sin(static_cast<int>(angle) * stepAngle) * r;
-    float y = radius + Cos(static_cast<int>(angle) * stepAngle) * r;
+    float x = static_cast<float>(radius) + Sin(static_cast<int>(angle) * stepAngle) * r;
+    float y = static_cast<float>(radius) + Cos(static_cast<int>(angle) * stepAngle) * r;
 
     dc.DrawCircle(static_cast<int>(x), static_cast<int>(y), radius / 5);
 }
@@ -73,13 +73,13 @@ bool GovernorGUI::MouseOnGovernor(wxMouseEvent &event) //-V2009
 
 float GovernorGUI::Sin(int grad)
 {
-    return sinf(grad * 3.1415926F / 180.0F);
+    return sinf(static_cast<float>(grad) * 3.1415926F / 180.0F);
 }
 
 
 float GovernorGUI::Cos(int grad)
 {
-    return cosf(grad * 3.1415926F / 180.0F);
+    return cosf(static_cast<float>(grad) * 3.1415926F / 180.0F);
 }
 
 
@@ -93,7 +93,7 @@ void GovernorGUI::OnTimer(wxTimerEvent &)
 
         if(delta != 0)
         {
-            angle += 0.1F * delta;
+            angle += 0.1F * static_cast<float>(delta);
 
             Refresh();
         }
